backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-genai==0.8.0
python-docx==1.1.2
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
python-multipart==0.0.9
sqlalchemy-utils
alembic==1.12.1
psycopg2-binary